// ProyectoAngular/PrimerAngular/src/app/models/profesores.ts
export class Profesores {
    id: string = "";
    cedula: string = "";
    correoelectronico: string = "";
    telefono: string = "";
    telefonocelular: string = "";
    fechanacimiento: string = "";
    sexo: string = "";
    direccion: string = "";
    nombre: string = "";
    apellidopaterno: string = "";
    apellidomaterno: string = "";
    nacionalidad: string = "";
    idCarreras: string = "";
    usuario: string = "";
}